requirements.txt
discord.py
requests
beautifulsoup4
pytube
mutagen
# logging
# urllib